Fix polygonize
